user usercredential change

+++ demoresourcemanagement/src/main/java/com/example/demoresourcemanagement/service/UserService.java
@@ -3,9 +3,13 @@ package com.example.demoresourcemanagement.service;
 import com.example.demoresourcemanagement.dao.UserDao;
 import com.example.demoresourcemanagement.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -17,17 +21,32 @@ public class UserService {
         userDao.save(user);
     }
 
-    public User getUserById(int id) {
-        return userDao.getOne(id);
+    public ResponseEntity<User> getUserById(int id) {
+        try{
+            Optional<User> user = userDao.findById(id);
+            return new ResponseEntity<>(user.get(), HttpStatus.OK);
+        } catch (NoSuchElementException e) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     public List<User> getAllUser() {return userDao.findAll();};
 
-    public void setUserById(User user) {
-        userDao.save(user);
+    public ResponseEntity<?> setUserById(int id, User user) {
+        Optional<User> existUser = userDao.findById(id);
+        if(existUser.isPresent()) {
+            userDao.save(user);
+            return new ResponseEntity<>(user, HttpStatus.OK);
+        }
+        return new ResponseEntity<>("User Not Found!", HttpStatus.NOT_FOUND);
     }
 
-    public void deleteUserById(int id) {
-        userDao.deleteById(id);
+    public ResponseEntity<?> deleteUserById(int id) {
+        Optional<User> existUser = userDao.findById(id);
+        if(existUser.isPresent()) {
+            userDao.deleteById(id);
+            return new ResponseEntity<>("Deleted User Successfully", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("User Not Found!", HttpStatus.NOT_FOUND);
     }
 }
